requirements.txt
anyio==3.7.0
certifi==2023.5.7
exceptiongroup==1.1.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
install==1.3.5
psycopg2-binary==2.9.6
python-telegram-bot==20.3
sniffio==1.3.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
